Add tuple and go styles
Examples in the readme.

I'm not entirely sure which api is the best, maybe version 1 can converge on a single style.

Or maybe it's best to give the options to the developers

'might-fail'
'might-fail/tuple'
'might-fail/go'

Again, this might change by version 1, but let's see what people do with this